# more small-array performance improvements