--- a/numpy/core/src/umath/ufunc_object.c
+++ b/numpy/core/src/umath/ufunc_object.c
@@ -2774,9 +2774,16 @@ reduce_type_resolver(PyUFuncObject *ufunc, PyArrayObject *arr,
      * resolution.
      */
     if (odtype != NULL) {
-        type_tup = PyTuple_Pack(3, odtype, odtype, Py_None);
-        if (type_tup == NULL) {
-            return -1;
+        if (!(PyTypeNum_ISBOOL(odtype->type_num) &&
+              PyArray_TYPE(arr) == odtype->type_num) &&
+            !(PyTypeNum_ISFLOAT(odtype->type_num) &&
+              PyArray_TYPE(arr) == odtype->type_num &&
+              PyArray_DESCR(arr)->byteorder == odtype->byteorder &&
+              odtype->byteorder == NPY_NATIVE)) {
+            type_tup = PyTuple_Pack(3, odtype, odtype, Py_None);
+            if (type_tup == NULL) {
+                return -1;
+            }
         }
     }
 
@@ -2784,7 +2791,7 @@ reduce_type_resolver(PyUFuncObject *ufunc, PyArrayObject *arr,
     retcode = ufunc->type_resolver(
                         ufunc, NPY_UNSAFE_CASTING,
                         op, type_tup, dtypes);
-    Py_DECREF(type_tup);
+    Py_XDECREF(type_tup);
     if (retcode == -1) {
         return -1;
     }