@juliantaylor 2015-10-09
as we are going to be less strict on preserving the ufunc structure this can be probably be considered for merging