@njsmith 2015-10-09
Maybe it would be best to finish off that NEP first and then figure out what to do here? Is this urgent?