@charris 2015-12-10
Ping folks who are in the conversation. We should either close this or take it forward.